refactor user page
